// src/app/share/model/product-detail.model.ts
export interface ProductDetailModel {
  createdDate: Date;

  description: string;

  id: number;

  idCategory: number;

  idSale: number;

  idMark: number;

  imageProduct: string;

  maSp: string;

  name: string;

  price: number;

  sellCount: number;

  totalItem: number;

  idColor: number;

  colorName: string;

  markName: string;

  categoryName: string;

  codeSale: number;

  descriptionMark: string;

  nameSale: string;
}